requirements.txt
streamlit>=1.34.0
matplotlib>=3.8.0
numpy>=1.26.0
fpdf2>=2.7.0
